Update patch set 1
Patch Set 1:

I'm trying to replace each heap reference load with a call to the read barrier entrypoint method. 

The current issue is that the heap reference load (Mir2Lir::LoadRefDisp() or Mir2Lir::LoadRefIndexed()) can happen in the middle of compiling a Dalvik instruction, so we need to save all temp registers, and restore them after the call. Saving registers can be done through Mir2Lir::FlushAllRegs(), but I couldn't find a way to restore them.

I read the code for GenInstanceof() and GenCheckCast(), but they don't need to restore the registers, as the calls are not in the middle of compiling an instruction. Is there any code that does something similar to what I'm trying to do? Or is there any easy way to restore the registers?

Patch-set: 1
Label: Verified=0