Update patch set 2
Patch Set 2:

> don't you do that just because DCE gets more precise during
 > RemoveDeadInstructions()?

Not really. It's because RemoveDeadBlocks does not update Phis, so they might be left with only one input (say when you remove one if-branch). Running phi elimination avoids duplicating code.

My point being, if some pass is an inseparable part of another optimization, why not let it manage it itself? What you're doing here is extracting logic from the inliner (and making a mess in the process).

Patch-set: 2