Update patch set 2
Patch Set 2:

> Some comment on the general algorithm:
 > Since we've already done loop analysis (essentially finding SCC's,
 > plus more like depth of loop. Also we only deal with natural
 > loops), it looks to me it can be simplified on iv detection side
 > with the loop info available.
 > 
 > Each iv must be a loop phi, or derived from another iv. To check
 > whether a loop phi is an iv, the input of the loop phi is traced
 > all the way back to itself, with only certain ops allowed. It's
 > like detecting SCC's but can be simpler to understand/implement. No
 > need for full Tarjan's algorithm. Does this do the same thing?

Thanks for your comments Mingyao. Note that loop detection finds SCCs in the control flow graph, this algorithm works on the SSA graph (a small but subtle difference). Using Tarjan over any ad-hoc scanning is essential, since it yields the very nice "on-demand" property that operands are already classified when classiying operations. With any ad-hoc method, one is bound to scan it incorrectly!

Patch-set: 2