Update patch set 2
Patch Set 2:

What happens if there are an odd number of characters in the string, and the character just past differs between the two strings.  Do we know that the string must be 0 terminated so that this is safe?  Otherwise we might say that the strings are different when they are identical.

Patch-set: 2
Reviewer: Gerrit User 1036869 <1036869@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0