Update patch set 6
Patch Set 6:

(9 comments)

I think it does not work because I modified the JDWP test to no longer use a breakpoint to suspend the program (the synchronization is done through a socket). Therefore, there is no longer any transition to interpreter during the test. Since the current CL asks for deopt only when leaving the interpreter, nothing happens.

Actually, we really need the debugger to install instrumentation exit stubs (for each method in the thread's stack) after setting a local variable in a Quick/Optimizing frame. This way, the instrumentation exit stub would trigger the deopt of the stack.

I'm taking a look at this now.

Patch-set: 6
Label: Verified=0