Update patch set 2
Patch Set 2:

I understand why you're doing this, but I'm worried about the engineering costs of always having to maintain that information when we add/change nodes. And since there are other ways to get the dex pc / line number information (and we do record that for important places such as instructions that throw or call sites), you should probably use those other ways instead.

The truly usefulness of this patch would also be for our own debugging when the generated code crashes. But as an ART developer that's pretty easy to compute.

Patch-set: 2