Update patch set 7
Patch Set 7:

(5 comments)

Thanks for updating. I successfully passed the test with your CL (PS6) by installing instrumentation exit stubs on all methods (return pc) of the thread's stack. This is now causing the deopt as expected.

However, currently I do this for all threads (not only the thread requested by the debugger) because the instrumentation is global in the runtime, not thread-local. Still I don't think we want other threads to pay the cost of executing instrumentation exit stub for nothing (because they won't need to deopt). So I'm now checking whether can do this only for one thread.

On the testing side, we also need to test the exception flow. Currently, the tests for StackFrame.Get/SetValues do not handle exception so I need to add the support.

Patch-set: 7
Label: Verified=0