Update patch set 1
Patch Set 1:

(1 comment)

As I say in my comment, StackVisitor::GetVReg has the same issue.


About testing, IMO it's going to be complicated to test that regression. From the code, the debugger will enventually call ShadowFrame::SetVReg to update the value then ShadowFrame::GetVReg to read it. And it will work... until the Object is moved by the GC at a different address. In that case, we'll have a stale reference in the ShadowFrame::vregs_ array and we'll crash. Indeed, ShadowFrame::SetVReg put null in the reference array so the GC does not even know the DEX register holds a reference, therefore it won't update it).

The test must trigger a GC that will move the reference (set by the debugger) so its pointer becomes invalid. The sequence would be:
1) Set a reference in DEX register vX with the debugger
2) Trigger a GC that moves that reference at another address
3) Read the reference from vX and use it (access a field, call a method). Without the CL, the runtime will crash on an invalid reference.

Patch-set: 1