Update patch set 2
Patch Set 2:

I would say it is just a clean-up.
It is a fix of the bug which probably is not reproducible now in AOSP-master. "Probably" is because I'm not aware about case in AOSP now which can lead to failure. The bug will be reproducible as soon as AOSP or someone else using AOSP will try to create HDivZeroCheck after SSA built. For example if you clone this instruction for any reason during the optimization you will get this bug.

So finally, it is not a clean-up but bug fix but I do not know whether the bug is reproducible now but it can easily be reproduced in the future at any time. According to above I think it makes sense to merge this bug fix.

Patch-set: 2