Update patch set 5
Patch Set 5:

> Catching up again... but looking at the code, the namespace is
 > making things ugly with all the duplication... Why did you/we go
 > with the namespace already?

We wanted some kind of isolation between generic and arch-specific nodes in the first place.  The first proposed solution was to use different files for arch-specific nodes, but concerns about future merging issues were raised, so the namespace solution was suggested.

It seems we have chosen the first solution, so maybe namespaces are superfluous now, except if we want some sort of prefix/suffix for arch-specific nodes.  E.g., instead of 

  class HArm64Foo { // …

we could have

  namespace arm64 {
    class HFoo { // …
  }

Or both:

  namespace arm64 {
    class HArm64Foo { …
  }  // namespace arm64

What do you think?

Patch-set: 5