Update patch set 3
Patch Set 3:

I'm ok with both options: arch:HFoo and HArchFoo. I don't think it makes a big difference as long as we always we the namespace prefix.

I'm more "concerned" how we handle this if we need to have the same instruction for let's say arm and arm64 (but not x86). I can think of MulAdd falling in this case.

Patch-set: 3
Label: Verified=0