Update patch set 8
Patch Set 8:

I see that my case was a lot simpler, involving a single pass.  I don't think that it would be terrible to have gvn.h included in a BE code generator, but I can see the other side as well.

Would it make sense for there to be a 'pass framework', where a BE could insert optimizations within the original optimizations, and/or replace an optimization with a BE specific one?

Patch-set: 8
Label: Verified=0