Update patch set 12
Patch Set 12:

> How about the codegen just returns the list of optimization passes
 > to optimizing_compiler.cc?

There's no difference between this and having conden_->RunArchOptimizations(). And I think the later is better as it skips one level of indirection.

I still think that having a codegen dependency on gvn,side_effects and such is undesirable and weird. 

Beside the dependency issue I think this blocks us having a more efficient pipeline. For example if we  run arch-simplifier before gvn (e.g. after fold2) we can save passes. With respect to this PS6 allows for greater flexibility.

Patch-set: 12