Update patch set 12
Patch Set 12:

> > > How about the codegen just returns the list of optimization
 > > passes
 > > > to optimizing_compiler.cc?
 > >
 > > There's no difference between this and having conden_->RunArchOptimizations().
 > > And I think the later is better as it skips one level of
 > > indirection.
 > >
 > > I still think that having a codegen dependency on
 > gvn,side_effects
 > > and such is undesirable and weird.

I am not a fan either, but the code is clean so I'm fine with it if it helps.

 > >
 > > Beside the dependency issue I think this blocks us having a more
 > > efficient pipeline. For example if we  run arch-simplifier before
 > > gvn (e.g. after fold2) we can save passes. With respect to this
 > PS6
 > > allows for greater flexibility.
 > 
 > It looks to me that getting to this efficient pipeline means each
 > architecture will decide on the optimizations it wants and just
 > return it to optimizing_compiler.cc. Is that what you think we
 > should do?

Do we expect different archs to run very different sets of passes in a different order?
For now I was only considering to run:

    instruction_simplifier_arch(split IRs)   // Example: array access split.
    GVN
    instruction_simplifier_arch(merge IRs)   // Example: muladd, add+shift

One drawback of interleaving arch-independent and arch-specific passes may be the checker tests. This can become a mess if we try to check a pass occurring after an arch-pass. Having a more flexible way of specifying what passes should run would help there.

Patch-set: 12
Label: Verified=0