Update patch set 4
Patch Set 4:

(4 comments)

Thanks all.

Calin: yes, we almost always successfully build the graph. Now that we're passing the values of parameters, a method can become much shorter than it originally is. So if a caller decides not to inline, it should not prevent other callers to try it.

The heuristic numbers I've put in were based on performance, with significant improvements on two benchmarks, namely Takl (x2) and ReversiBench (x1.5).

They do have a cost on compile-time and code size though: 10% compile-time, 5% code size.

I'll change the values to be more inlined with what we had before. And file a bug for better inlining for Takl and ReversiBench.

Patch-set: 4
Label: Verified=0