Update patch set 5
Patch Set 5:

> > It would be interested how much of the gain is attributed to the
 > > two dimensions:
 > > a) the parameter handling
 > > b) the fact that we allow per-caller inlining.
 > >
 > > 10% sounds like a lot, but the gains are pretty good as well.
 > 
 > Not sure I see the difference between 1) and 2) ?

Folding the param reduces the size of the graph allowing more to be inlined. But I was wondering, if we keep the check (if we fail to inline once don't bother again). Do we still get a relevant speed-up? i.e. how much just adding the dce and fold over the generic graph helps?

Patch-set: 5