Update patch set 7
Patch Set 7:

> (1 comment)
 > 
 > LGTM. Are there any compelling reasons why we should go for the
 > dlopen version instead?

dlopen could allow people to develop & deploy new versions of backends without having to recompile libart-compiler.so. With additional work it might be possible to have entirely new backends be deployed in this way (I think that would need reworking in how we identify isa's though).

On the flipside doing this is far more complicated, it might make changing the backend api more difficult, and its not clear how useful/used this feature would be.

Patch-set: 7