Update patch set 1
Patch Set 1:

> Yuck. Partial support is pretty weird and seems just targeted at
 > bogus dex obfuscator/manipulator.

It's not bogus, but yes, that's targetted at that one obfuscator. (Well, the unlock fix isn't...)

 > 
 > Verifier should follow (our absent) specification. If we want this
 > change, we should really debater whether we would write a
 > specification that says "except for immediately following moves,
 > lock aliasing is forbidden"?

Lock aliasing is already allowed, and tracked (slightly incorrectly). If you do the move-object *after* the monitor-enter, things (mostly) work.

The problem is a non-low-overhead solution for register aliasing established *before* the monitor-enter. Yes, we need a better solution, but currently I don't know a good solution that's not screwing performance and memory requirements royally.

I'm generally more interested in getting balanced working good enough that we can turn it on. For that, I'm thinking of an alternative approach, anyways. But this is a small enough change that helps us right now...

Patch-set: 1
Label: Verified=0