Update patch set 2
Patch Set 2:

I'm glad this was done, all the explicits actually hurt because it makes it incompatible with list-init.

i.e.  if we had this code:

  explicit Foo(int x, int y);
  Foo f({1, 2}); // is a compiler error

Adding explicit to a copy/move constructor is even worse (as we discovered with agampe) since implicit conversion is used to do returns or pass arguments to function. Another thing that should probably not happen.

  explicit Foo(Foo&& other);
  Foo create() {
    return Foo{}; // compiler error
  }

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0