Update patch set 3
Patch Set 3: Verified-1

FYI, here is the corresponding test: https://android-review.googlesource.com/167782

However, it catches another bug:
  Check failed: PcIsWithinQuickCode(reinterpret_cast<uintptr_t>(code), pc) void org.apache.harmony.jpda.tests.jdwp.Events.SingleStepWithPendingExceptionDebuggee.run() pc=2ba78c3f152f code=0x2ba78c3f168c size=1d6

This happens after we inserted a breakpoint and resume the debuggee. This is an assert durint stack walk. I'm investigating to know if it is caused by my change.

Patch-set: 3
Label: Verified=-1