Update patch set 1
Patch Set 1:

Although it is nice to see the combined effect, would you mind splitting this CL into the utils change and, as follow up, use in optimizing? That way we get two nice small CLs.

Also, please add a test in the first CL on the new instructions and short label example.

Patch-set: 1
Label: Code-Review=0