Update patch set 2
Patch Set 2:

> > > Why did the base register change from %eax to %al between PS1
 > and
 > > > PS2? This doesn't compile for aosp_x86_64-eng.
 > >
 > > %al forced it to use the actual tight encoding, whereas %eax
 > always
 > > produced a rex prefix. I personally don't understand why the
 > Clang
 > > assembler puts down the rex prefix.
 > >
 > > What's the error you see (I assume gnu-as)?
 > 
 > Error: `(%al)' is not a valid base/index expression
 > 
 > Not sure what assembler - the AOSP default.

Pretty sure we haven't switched x86-64 to Clang, so it would be GCC's. OK, I'll have to find another version. I tried the "write %ah to memory," and that had a consistent result, too (because you can't address %ah with rex prefixes), so I'll see what GCC says to it.

Patch-set: 2