Update patch set 3
Patch Set 3:

I talked with Mathieu offline.

It sounds like the issue is that a DexFile shouldn't be deleted while its corresponding DexCache is still alive. The fix should be to avoid repeatedly opening DexFile objects in oatdump, and ensure the DexFile objects that are opened stay alive as long as the class linker is alive. Ideally all the dex files needed by oatdump are opened eagerly and reused. A temporary workaround is to cache and reuse the opened dex files in oatdump. I'm fine with that approach. My main concern with PS1 through PS3 is we shouldn't expose the class linker's opened_dex_files.

Patch-set: 3