Update patch set 1
Patch Set 1:

> I don't think this deserves a verbosity class. This is compiler
 > internals, and we're normally not exposing those more than
 > verbose:compiler. Why not log in debug mode?

We know that inliner is very heavy optimization as for now and this new verbosity class would allow us to get what we inline/reject now and how to improve the inliner. In a debug mode compiler prints too much information, so this makes it hard to separate the messages from inliner.

Patch-set: 1