Update patch set 3
Patch Set 3: Code-Review+1

This isn't really testing compaction at all, this is testing moving, so only +1.

You should be allocating a sequence of objects, "free" some of them, and then see that the result is compact. That's possible with a visitor. See, for example, https://android-review.googlesource.com/93285.

Patch-set: 3
Label: Code-Review=+1