Update patch set 2
Patch Set 2:

(1 comment)

> (2 comments)
 > 
 > Why do we need to affect the ShadowFrame itself?

I want to add a field that potentially allocates (a pointer to a map or vector) for lock counting. I do not want callers to know about this and have to explicitly call the destructor.

Right now we're relying on the ShadowFrame not having any heap-referencing fields, so when the alloca dies the shadow frame is fully gone. That wouldn't be true anymore...

Patch-set: 2