Update patch set 1
Patch Set 1: Code-Review-1

Do we want to keep this?

I somehow feel that we shouldn't really support debug symbols for the JIT. If people want to do native/performance debugging, use an AoT build.

So I'd vote for removing the gdb jit hook completely, considering it was never adapted to the mcjit (it triggers on opening an ELF file).

Patch-set: 1