Update patch set 2
Patch Set 2:

> Does the testcase cause deoptimization even though it doesn't throw
 > ArrayIndexOutOfBoundException?
 > 
 > If it throws AIOOBE, currently the method that can catch that
 > exception is compiled by quick so optimizing compiler doesn't play
 > a role there and we are safe.
 > 
 > However it is indeed fragile. Due to code motion, liveness in
 > interpreter is different from liveness in compiled code, so any env
 > should probably just hold values live, in additional to invoke.

In the test code, foo1 and foo3 are compiled by Optimizing and foo2 is done by Quick (foo2 has try/catch block).
The expected behavior is the following.
foo3 throws AIOOBE whenever it is invoked by foo2 (k+1000 is always greater than the length of array b[]).
Deoptimization is generated right before the loop body of foo3 and it is always executed when foo3 is invoked.
When deoptimization is executed, foo3 enters interpreter mode. 
When interpreter executes the bytecode of foo3, AIOOBE is thrown and caught by foo2 (catch block does nothing).
Returning from foo2, the loop body of foo1 continues in interpreter mode.
This process is repeated and the program is finished showing the message 'finish'.
Without this patch, however, the behavior is not like the above.
When a[k][k] is accessed returning from foo2, AIOOBE is thrwon and the program is finished on the second iteration of the loop of foo1.

Patch-set: 2
Label: Verified=0