Update patch set 7
Patch Set 7:

(2 comments)

Is there ever a time we can make use of a corrupt or invalid image file? Could we treat it the same as a missing image file (with some vlogging in the image assistant to note the distinction if desired)? Can you summarize all of the relevant image states with two booleans:
  1. We have or don't have an uncorrupted system image.
  2. We have or don't have an uncorrupted cached image. 

That would be much easier for me to understand than the current ImageState + ImageInfo combination.

Patch-set: 7