Update patch set 4
Patch Set 4:

I don't have any preference towards one style or another. 

I do like "for ( : ) " syntax as it makes things shorter but that's about the only thing. (the rest of the api is slightly worse imo)

Besides the cosmetics, the only functional change of the CL is the ability to track object allocation. This is nice and needed to get better understanding of our memory use. So unless we have other means to do it I'd go forward with the CL.

And just to verify my assumption, there's no performance difference, right?

Patch-set: 4
Label: Verified=0