Update patch set 4
Patch Set 4:

> And just to verify my assumption, there's no performance
 > difference, right?

The index-based loop has a chance of missed optimizations where the compiler must reload the GrowableArray::elem_list_ because it cannot prove that it didn't change. So this CL may improve performance but I would expect the difference to be insignificant and lost in the noise.

Patch-set: 4
Label: Verified=0