Update patch set 3
Patch Set 3: Verified-1

There is one issue left: we report the exception twice to the debugger now :(   (the first time from the QuickExceptionHandler and the second time from the interpreter when executing the first deoptimized frame).

I need to find a way to avoid the double notification. I think the interpreter should just not reporting the exception event for the first deoptimized frame because it has already been reported by the QuickExceptionhandler, just before we deoptimize the stack.

This is tested by https://android-review.googlesource.com/169560.

Patch-set: 3
Label: Verified=-1