Update patch set 3
Patch Set 3:

> > There is one issue left: we report the exception twice to the
 > > debugger now :(   (the first time from the QuickExceptionHandler
 > > and the second time from the interpreter when executing the first
 > > deoptimized frame).
 > >
 > > I need to find a way to avoid the double notification. I think
 > the
 > > interpreter should just not reporting the exception event for the
 > > first deoptimized frame because it has already been reported by
 > the
 > > QuickExceptionhandler, just before we deoptimize the stack.
 > >
 > > This is tested by https://android-review.googlesource.com/169560.
 > 
 > Did you give any thought to the approach I mentioned in my first
 > comment, i.e. punt the whole thing to interpreter when
 > IsForcedInterpreterNeededForException() is true? Let the
 > interpreter do FindCatch. I don't think we need to optimize
 > performance that much for handling throwing exception in debugger.

This will not work if debugger needs deoptimization *after* being notified of the exception, not *before*. The notification happens in FindCatch. Checking IsForcedInterpreterNeededForException before will only work for debugging events set *before* the exception happens, thus you will miss events set *after* the exception is thrown.

When you suspend your program on an EXCEPTION event (at the time of the throw), you can request a BREAKPOINT (or any other event) in the method catching this exception. Therefore, you need to perform a deoptimization to execute that method with the interpreter. (This is the base of new tests in https://android-review.googlesource.com/169560)

Patch-set: 3