Update patch set 1
Patch Set 1:

The do_binary_search is a simple 47 line perl script that just calls the script 'execute' with the current number and then tests the return code for success.  It is simple to write the 'execute' script so that it passes that number to a parameter to dex2oat and then does whatever is necessary to see if it is successful.

I will gladly donate this script to whoever asks.

Patch-set: 1