Update patch set 13
Patch Set 13:

I tried my best to test this. The problem is that our run-tests don't work, as the tests are compiled against the standard Java libraries which don't have the specialized version.

The (AOSP) emulator boots fine, so I assume it's fine. Gotta see if fugu breaks down the line.

Patch-set: 13