Update patch set 1
Patch Set 1:

This is just a more refined breakdown of why we reject some classes in the compiler. It makes it easier to track the route cause, without having to enable verbose output for the verifier.
It basically expands the previous kNotCompiledVerificationFailed into it's subcategories.

So, I still think it belongs to the compiler. However, as we discuss offline, I can keep it in my private tree until i finish with the unresolved support.

Patch-set: 1
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0