Update patch set 8
Patch Set 8:

> My previous comments focused on the image assistant API and how it
 > cleans up various image code scattered across the code base in
 > patchoat, image space, VMRuntime, runtime, and the oat file
 > assistant. From that perspective, I see minor improvements to code
 > clarity where uses of the image assistant have replaced the
 > existing code, but I don't think those minor improvements are worth
 > the added complexity of the new image assistant interface.
 > 
 > In our discussions, Alex suggested the primary value of the image
 > assistant was to improve the image_space code. When I focus on the
 > image_space code, it does not appear to me to be cleaner or easier
 > to understand, aside from moving some code out into a separate
 > source file. To me, the benefit the image assistant has for the
 > image_space code is not worth the added complexity of the image
 > assistant API.


Seriously?

This code replaces a 5-level deep, 100+ line long if-else tree that uses the output of a helper function that returns 7 (!) different values to determine which of 3 (relatively simple) actions to take. The new version is a 3 case switch statement where each of the cases have descriptive names and do exactly one thing. If this doesn't count as cleaner or easier to understand I really don't know what would.

As for ImageInfo/State being rather complicated: The fact is that it needs to encapsulate a decision tree that has 12 'if-else' statements in it going up to 5 levels deep to be used by the image_space (in more then one place). Of course it will have a fair number of possible values and a rather complicated process associated with creating it.


 > I think this CL is trying to do too much all at once. It will be
 > easier to see the value in smaller, more focused cleanup CLs. For
 > example:
 > 1. A CL to clean up the image_space code


The fact is that any reasonable cleanup will end up looking a hell of a lot like image_assistant and ImageState because they encapsulate the decision tree that needs to be cleaned up.


 > 2. A CL to factor out duplicate logic in patchoat, VMRuntime, runtime, and the oat file assistant.


Into what? I put it into the image_assistant because that would need all that logic anyway. If we don't put it there we will still have duplicate logic because the fact is that the only thing these have in common with each other is they all duplicate different parts of logic used in image_space (and now image_assistant).


 > 3. A CL to remove unused IsImagePic code from patchoat.


(A) This is an absurdly trivial change. (B) I only included this because you asked me to.


 > 4. A CL to fix the bug(?) in common_runtime_tests.

Patch-set: 8
Label: Verified=0