Update patch set 3
Patch Set 3:

I implemented the same thing but have not pushed it for review. I was waiting for previous patches to go in. But I'm happy for this one to go forward.

Can we make the new IR a 'common' IR? We will do the same thing on arm64. What about `HConditionalSelect`?

One further optimization I was thinking about was to perform the optimization on non-empty 'true' and 'false' blocks. For example when the 'true' or/and 'false' block contain a small number (1? 2? it requires profiling) of 'fast' IRs without side-effects (eg. simple binops).

Patch-set: 3
Label: Verified=0