Update patch set 4
Patch Set 4:

> > > Seems like overkill
 > >
 > > How overkill?
 > 
 > You fully stop a function from being inlined.

Sure, but considering how big it is, being in the -inl.h file seemed artificial.

 > 
 > >
 > > > to move the whole function, considering it was
 > > > in the inl file to begin with. How about writing a function
 > that
 > > > returns the exit stub address (that can go into the .cc file),
 > > and
 > > > caching it in a local variable?
 > >
 > > What do you win from doing this?
 > 
 > That the main body is still inlined at call sites, which seemed to
 > have been the intent, considering that it was in an -inl file.

Initially the method was 2 lines. I added the inlining support afterwards. Now that we need an extra check, method is big enough that inlining it is maybe not worth it.

Patch-set: 4
Label: Verified=0