Update patch set 5
Patch Set 5:

(1 comment)

Should only perform the transformation if the `HCondition` has only one use?

The transformation can be 'bad' when the `HCondition` has multiple uses and needs to be materialized as it extends the live ranges of the two values to compare. In that situation it would be better to use the `HCondition` as an input to the `HConditionalSelect`.

The problem is that we don't know until `PrepareForRegisterAllocation` has run whether the condition needs materialization.

It looks like `NeedsMaterialization()` is only used in the codegen. Maybe instead of setting the value at reg allocation time we could evaluate it dynamically?

Example code (not tested)

    boolean cond = x == y;
    int force_materialisation = input + (int)cond;

    int a = cond ? a1 : a2;
    int b = cond ? b1 : b2;

    if (cond) {
      d = d1;
    } else {
      d = d2;
    }
    if (cond) {
      e = e1;
    } else {
      e = e2;
    }

Digressing:
If we remove the use forcing materialisation maybe doing the transformation everywhere is still worth it.
Ideally we would generate something like

    cmp x, y
    csel.eq a, a1, a2
    csel.eq b, b1, b2

(or do some basic block merging and use branches).
Maybe we can do that if we can express condition flags production/consumption on the IRs, but I can see very big hurdles on the way.

Patch-set: 5
Reviewer: Gerrit User 1052304 <1052304@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0