Update patch set 20
Patch Set 20:

(11 comments)

(Not fully done, but want to give incremental feedback)

Too few reviewers. This code involves class-linker, so Mathieu should be reviewer. And the GC things ask for Hiroshi.

 > 
 > * vector of mirror::Object* in interruptible code is unsafe (thanks
 > for spotting, Nicolas) for GCs. best way to fix this?

Yeah, Hiroshi's gonna be mad. :-)

 > * preverified flag being cleared on default methods, resulting in
 > potentially duplicated verifications (agampe, thoughts?)

No, that's fine. The verifier only verifiers the methods from the class-def, not the Class.

 > * dex file test should get updated so non-abstract methods are
 > allowed in interface method flags (agampe, thoughts?)

Of course.

Patch-set: 20