Update patch set 2
Patch Set 2:

> > > cause now buildbot-build.sh doesn't work out of the box (as I
 > > would
 > > > expect) without setting the env var.
 > >
 > > As *you* would expect, but it's not its purpose anymore :-)
 > 
 > You still end up with a script that does a partial job (relying on
 > "implicit" env vars. If someone else want to look on the buildbots
 > and make changes this will not be obvious and waste some time...

The buildbot lists the env vars when executing a command. Once I restart it, you'll see:
CUSTOM_TARGET_LINKER=....

I'm still unclear on what your issue is. The script is now just building what the buildbot needs. The fact that you (and I) relied on it setting the linker is an old behavior that does *not* work anymore with the ninja change.

Patch-set: 2