Update patch set 9
Patch Set 9:

(2 comments)

> (2 comments)
 > 
 > Correct?

More or less.

 > We enable lock counting unconditionally all the time (as opposed to
 > certain verification runs)

You mean we, the runtime as a whole, enforce structured locking all the time? Yes, that's true.

 > but limited to methods that failed the verification

Yes. If the verifier can't guarantee balanced locking (a stronger constraint than structured locking), or can show that structured locking is violated, a method will soft-fail verification. In that case, the method will, just like for all the other failures, be run in the interpreter in the access_checks mode. The access_checks mode now also registers lock and unlock operations (note that my commit message incorrectly refers to the verifier).

 > and the reason that we don't need to do this in the
 > compiler is that we won't compile such methods to begin with.

Yes, the compilers will not compile code that the verifier failed because of locking issues. The assumption is that most code we'll see is well-formed.

Patch-set: 9
Label: Verified=0