Update patch set 2
Patch Set 2:

> Not sure how best to test this. Could we have an instrumentation
 > test that wouldn't require --debuggable?

IIRC only the debugger is using instrumentation stubs atm but the JDWP tests run with the --debuggable flag. The tracing feature uses the interpreter so the --debuggable flag wouldn't change anything when running run-tests with tracing active.

An alternative may be to add a JNI-like run-test where we could instrument the code with or without the --debuggable flag.

Patch-set: 2