Update patch set 2
Patch Set 2:

This CL shows how to handle the TwoPath case. We have to relax the failure detection on merge cases.

This is maybe not fully correct as-is. It trades a too lax case for slightly better checking, and relaxes a too-strict check to something that I'm not sure about full correctness.

What this patch does is the following: assume you're merging v0 in a join point, which was an alias of v1 in one path, but v2 in another. Then the lock levels don't match, and we rejected. But it's OK, because the locking info is still there - in v1 and v2 - so we'll still complain about missing unlocks if they are retained till a return point.

Where I am not sure: while an empty lock state doesn't allow unlocks of v0 afterwards, it does allow new locks. So a monitor-enter afterwards isn't rejected, even though we lost all alias info and previous lock state (so it's not precise anymore), while we don't support sets of lock state.

I *believe* that is OK, it still forces you to explicitly unlock on v0 *in addition* to the locks from the former aliases.

Patch-set: 2