Update patch set 5
Patch Set 5:

(2 comments)

> (2 comments)
 > 
 > I still think it's cleaner to generalize dlmalloc_space to handle
 > jit's and gc's use cases. Add a TODO?

Alternatively is it worth refactoring dlmalloc_space? As you're saying it's very Java Object specific, and now that I played a bit with it, there is probably as much code working around it compared to using mspace and allocating bitmaps manually. WDYT? I could do that in a follow-up change.

Patch-set: 5