Update patch set 2
Patch Set 2:

(4 comments)

Are you sure the stack map agree with the vision the interpreter (thus the verifier) would have about the DEX registers?

I mean that a DEX register won't be assigned to kDeadValue because the compiler optimized it while the interpreter is going to use it.

Patch-set: 2
Label: Verified=0