Update patch set 3
Patch Set 3:

(1 comment)

>> 1) Changing allocations to use Scoped arenas
>> 2) Changing strings to string piece
> 1 and 2 are the same.
>> 3) Algorithmic changes
> The algorithmic changes are using different std data structures which is not complex.

Mathieu, the point is that, like Andreas was saying, you have a lot of separate changes in this CL right?

I roughly subdivided it into 3 separate areas (#3 could be further separated). Andreas' question was if we profile these changes separately, where is the real performance win?

It's possible that only #3 brings the majority of the 22% speedup we are seeing (or maybe it's only #1+#3, or maybe #2+#3). The point is, the more data the better.

Plus it's a valuable data point for brand new code we write, if this is really more efficient.

Patch-set: 3
Label: Verified=0