Update patch set 1
Patch Set 1:

So jdwp could be fine, as it takes around 10 minutes for both target and host. libcore is more problematic, as it's 30 minutes on host and at most 2 hours on target.

We could move any non-JIT testing to FYI bots, though that will take some time to setup.

Note that with this change, we still do test AOT through the run-tests.

Patch-set: 1