Update patch set 2
Patch Set 2:

Spoke with nnk about this. No files should be written to /data/local/tmp. It still exists for backwards compatibility since some third party apps rely on its existence. Let's move the file write elsewhere. Since this is written from an app's process can we write the trace file to the app's home directory?

Example of mischief that could occur as-is: the shell uid could create a symlink here to the traced app's home directory overwriting arbitrary files with the traced data. We need to set this up so that an app's data is protected from such corruption.

Patch-set: 2
Label: Code-Review=0